Add some tests of the comments view (#173165)
* Add some tests of the comments view

* Fix tests
