+++ src/vs/workbench/contrib/comments/browser/commentsView.ts
@@ -11,7 +11,6 @@ import { isCodeEditor, isDiffEditor } from 'vs/editor/browser/editorBrowser';
 import { IInstantiationService, ServicesAccessor } from 'vs/platform/instantiation/common/instantiation';
 import { IThemeService } from 'vs/platform/theme/common/themeService';
 import { CommentNode, CommentsModel, ResourceWithCommentThreads, ICommentThreadChangedEvent } from 'vs/workbench/contrib/comments/common/commentModel';
-import { CommentController } from 'vs/workbench/contrib/comments/browser/commentsEditorContribution';
 import { IWorkspaceCommentThreadsEvent, ICommentService } from 'vs/workbench/contrib/comments/browser/commentService';
 import { IEditorService, ACTIVE_GROUP, SIDE_GROUP } from 'vs/workbench/services/editor/common/editorService';
 import { CommandsRegistry } from 'vs/platform/commands/common/commands';
@@ -41,6 +40,7 @@ import { CommentThreadState } from 'vs/editor/common/languages';
 import { IDisposable, MutableDisposable } from 'vs/base/common/lifecycle';
 import { ITreeElement } from 'vs/base/browser/ui/tree/tree';
 import { Iterable } from 'vs/base/common/iterator';
+import { CommentController } from 'vs/workbench/contrib/comments/browser/commentsController';
 
 const CONTEXT_KEY_HAS_COMMENTS = new RawContextKey<boolean>('commentsView.hasComments', false);
 const CONTEXT_KEY_SOME_COMMENTS_EXPANDED = new RawContextKey<boolean>('commentsView.someCommentsExpanded', false);
@@ -464,7 +464,20 @@ export class CommentsPanel extends FilterViewPane implements ICommentsView {
 		this.someCommentsExpandedContextKey.set(this.isSomeCommentsExpanded());
 	}
 
-	private isSomeCommentsExpanded(): boolean {
+	public areAllCommentsExpanded(): boolean {
+		if (!this.tree) {
+			return false;
+		}
+		const navigator = this.tree.navigate();
+		while (navigator.next()) {
+			if (this.tree.isCollapsed(navigator.current())) {
+				return false;
+			}
+		}
+		return true;
+	}
+
+	public isSomeCommentsExpanded(): boolean {
 		if (!this.tree) {
 			return false;
 		}
